Move theta rho geometry operations into geometry.js
