+++ src/common/geometry.js
@@ -138,3 +138,78 @@ export const circle = (radius=1, start=0) => {
 
   return points
 }
+
+// Subsample lines into smaller line segments
+export const subsample = (vertices, maxLength) => {
+  let subsampledVertices = []
+  let previous = undefined
+  let next
+
+  for (next = 0; next < vertices.length; next++) {
+    if (previous !== undefined) {
+      const start = Victor.fromObject(vertices[previous])
+      const end = Victor.fromObject(vertices[next])
+
+      const delta = end.clone().subtract(start)
+      const deltaSegment = end.clone().subtract(start).normalize().multiply(Victor(maxLength, maxLength))
+
+      // This loads up (start, end].
+      for (let step = 0; step < (delta.magnitude() / maxLength) ; step++) {
+        subsampledVertices.push(
+          new Victor(start.x + step * deltaSegment.x,
+                     start.y + step * deltaSegment.y))
+      }
+
+    }
+    previous = next
+  }
+
+  // Add in the end
+  if (previous !== undefined) {
+    subsampledVertices.push(vertices[vertices.length - 1])
+  }
+
+  return subsampledVertices
+}
+
+// Convert x,y vertices to theta, rho vertices
+export const toThetaRho = (subsampledVertices, maxRadius, rhoMax) => {
+
+  let vertices = []
+  let previousTheta = 0
+  let previousRawTheta = 0
+
+  // Normalize the radius
+  if (rhoMax < 0) { rhoMax = 0.1 }
+  if ( rhoMax > 1) { rhoMax = 1.0 }
+
+  for (let next = 0; next < subsampledVertices.length; ++next) {
+
+    const rho = (Victor.fromObject(subsampledVertices[next]).magnitude() / maxRadius) * rhoMax
+
+    // What is the basic theta for this point?
+    let rawTheta = Math.atan2(subsampledVertices[next].x,
+                              subsampledVertices[next].y)
+    // Convert to [0, 2pi]
+    rawTheta = (rawTheta + 2.0 * Math.PI) % (2.0 * Math.PI)
+
+    // Compute the difference to the last point.
+    let deltaTheta = rawTheta - previousRawTheta
+
+    // Convert to [-pi,pi]
+    if (deltaTheta < -Math.PI) {
+      deltaTheta += 2.0 * Math.PI
+    }
+    if (deltaTheta > Math.PI) {
+      deltaTheta -= 2.0 * Math.PI
+    }
+
+    const theta = previousTheta + deltaTheta
+
+    previousRawTheta = rawTheta
+    previousTheta = theta
+    vertices.push(new Victor(theta, rho))
+  }
+
+  return vertices
+}
